agent/scripts/metadata_parser.py
"""
metadata_parser.py

Placeholder script that parses local X-ray data. In practice, you'd:
  - Scan a folder for new files (e.g., DICOM, CSV logs, etc.).
  - Extract metadata (timestamps, specimen ID, etc.).
  - Output structured metadata (JSON, YAML, etc.) for further processing.
"""

import os
import yaml
import sys

def main():
    print("Running metadata_parser placeholder...")
    # Example usage: python3 metadata_parser.py /home/pi/xray_data
    # Parse input arg for the data directory
    if len(sys.argv) < 2:
        print("Usage: python3 metadata_parser.py <data_directory>")
        sys.exit(1)

    data_dir = sys.argv[1]
    # TODO: Implement actual parsing logic
    sample_metadata = {
        "specimen_id": "ABC123",
        "scan_date": "2024-12-28T10:15:00Z",
        "comments": "Placeholder from metadata_parser.py"
    }

    output_file = os.path.join(data_dir, "parsed_metadata.yaml")
    with open(output_file, 'w') as f:
        yaml.dump(sample_metadata, f)
    print(f"Metadata parser wrote placeholder data to {output_file}")

if __name__ == "__main__":
    main()
